Handle controller and views within ConnectFour
