+++ config/connect-four/routing.yml
@@ -1,0 +1,15 @@
+connect_four_open:
+    path: /api/connect-four/games/open
+    methods: [POST]
+    controller: connect-four.page-controller::openAction
+
+connect_four_abort_challenge:
+    path: /api/connect-four/challenges/{gameId}/abort
+    methods: [POST]
+    controller: connect-four.page-controller::abortChallengeAction
+    defaults: { _format: json }
+
+connect_four_challenge:
+    path: /challenge/{id}
+    methods: [GET]
+    controller: connect-four.page-controller::challengeAction
