+++ config/connect-four/services/controller.yml
@@ -4,3 +4,9 @@ services:
         arguments: ['@connect-four.query-bus']
         calls: [[setContainer, ['@Psr\Container\ContainerInterface']]]
         tags: ['controller.service_arguments', 'container.service_subscriber']
+
+    connect-four.page-controller:
+        class: Gaming\ConnectFour\Port\Adapter\Http\PageController
+        arguments: ['@connect-four.command-bus', '@connect-four.query-bus', '@web-interface.security']
+        calls: [[setContainer, ['@Psr\Container\ContainerInterface']]]
+        tags: ['controller.service_arguments', 'container.service_subscriber']
