+++ config/web-interface/services/controller.yml
@@ -1,9 +1,8 @@
 services:
     web-interface.page-controller:
         class: Gaming\WebInterface\Presentation\Http\PageController
-        arguments: ['@connect-four.query-bus']
-        calls: [[setContainer, ['@Psr\Container\ContainerInterface']]]
-        tags: ['controller.service_arguments', 'container.service_subscriber']
+        arguments: ['@twig', '@connect-four.query-bus']
+        tags: ['controller.service_arguments']
 
     web-interface.signup-controller:
         class: Gaming\WebInterface\Presentation\Http\SignupController
@@ -34,5 +33,4 @@ services:
     web-interface.connect-four-controller:
         class: Gaming\WebInterface\Presentation\Http\ConnectFourController
         arguments: ['@connect-four.command-bus', '@web-interface.security']
-        calls: [[setContainer, ['@Psr\Container\ContainerInterface']]]
-        tags: ['controller.service_arguments', 'container.service_subscriber']
+        tags: ['controller.service_arguments']
