+++ src/ConnectFour/Port/Adapter/Http/View/open-game.html.twig
@@ -1,0 +1,15 @@
+<div class="dropdown">
+    <input type="checkbox" class="gp-dropdown-toggle" id="open-game-dropdown">
+    <label class="btn btn-primary w-100 dropdown-toggle"
+           for="open-game-dropdown">
+        Open a new game
+    </label>
+    <div class="dropdown-menu dropdown-menu-card dropdown-menu-arrow dropdown-menu-end"
+         data-bs-popper>
+        <div class="card">
+            <div class="card-body">
+                {{ form(form, {action: path('connect_four_open')}) }}
+            </div>
+        </div>
+    </div>
+</div>
