+++ src/WebInterface/Presentation/Http/View/lobby.html.twig
@@ -17,21 +17,7 @@
         <div class="col-lg-3 order-first order-lg-last">
             <div class="card">
                 <div class="card-body">
-                    <div class="dropdown">
-                        <input type="checkbox" class="gp-dropdown-toggle" id="open-game-dropdown">
-                        <label class="btn btn-primary w-100 dropdown-toggle"
-                               for="open-game-dropdown">
-                            Open a new game
-                        </label>
-                        <div class="dropdown-menu dropdown-menu-card dropdown-menu-arrow dropdown-menu-end"
-                             data-bs-popper>
-                            <div class="card">
-                                <div class="card-body">
-                                    {{ form(openForm, {action: path('open')}) }}
-                                </div>
-                            </div>
-                        </div>
-                    </div>
+                    {{ render_ssi(controller('connect-four.fragment-controller::openAction')) }}
                 </div>
             </div>
         </div>
